Added skeleton indicator component
